% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablerdemo-package.R
\docType{package}
\name{tablerdemo-package}
\alias{tablerdemo}
\alias{tablerdemo-package}
\title{tablerdemo: ADD TITLE}
\description{
ADD DESCRIPTION. TWO OR MORE LINES
}
\seealso{
Useful links:
\itemize{
  \item \url{https://WEBSITE.COM}
  \item Report bugs at \url{https://github.com/USERNAME/PKGNAME/issues}
}

}
\author{
\strong{Maintainer}: YOUR NAME \email{YOUR@EMAIL.COM} (\href{https://orcid.org/0000-0001-0002-0003}{ORCID})

Authors:
\itemize{
  \item YOUR NAME \email{YOUR@EMAIL.COM} (\href{https://orcid.org/0000-0001-0002-0003}{ORCID})
}

}
\keyword{internal}
